+++ src/security-liquibase/src/main/resources/properties/liquibase-local.properties
@@ -1,5 +1,5 @@
 url=jdbc:mysql://localhost/security_sample
-username=root
-password=root
+username=sec_user
+password=sec_user
 logLevel=finest
 port=8080
